http: Fix render err/panic when context times out
This fix removes the possibility to write on the response writer from different threads. That prevents renderers from writing a response body when the response writer stream has closed due to a cancellation or timeout.

This commit fixes the "http: wrote more than the declared Content-Length” error.
It also fixes panics when a race condition occurs on the response stream.